Implement support for "Named Webhooks" (#62)
Webhooks are now stored in a database table, and the JWT implementation is removed, even for ephemeral webhooks.

A new flavor of webhook is added here dubbed "named webhooks". A named webhook specifies the `url` property when being created, and the user is expected to define the API route in their framework - in which the `handleWebhook()` function is expected to be invoked in order for the relevant workflow runs to be resumed upon receiving a request to the webhook.

Below is an example of initiating a background chat request from OpenAI, and then waiting on the webhook event to be notified of that chat's completion:

```ts
import { useWebhook } from '@vercel/workflow-core';
import OpenAI from 'openai';
import { z } from 'zod';

const openai = new OpenAI();

async function initiateOpenAIResponse() {
  'use step';
  const resp = await openai.responses.create({
    model: 'o3',
    input: 'Write a very long novel about otters in space.',
    background: true,
  });
  console.log('OpenAI response:', resp);
  return resp.id;
}

async function getOpenAIResponse(respId: string): Promise<string> {
  'use step';
  const resp = await openai.responses.retrieve(respId);
  return resp.output_text;
}

export async function openAiBackgroundWebhookWorkflow() {
  'use workflow';

  // Initiate a background "Response" request to OpenAI,
  // which will invoke the webhook when it's done.
  const respId = await initiateOpenAIResponse();

  // Register the webhook URL and provide the schema for the
  // type of events that we are interested in, and specifically
  // for the exact response ID that we just created.
  const webhook = useWebhook({
    url: '/api/openai/webhook',
    body: z.object({
      type: z.string().startsWith('response.'),
      data: z.object({
        id: z.literal(respId),
      }),
    }),
  });
  console.log('Registered webhook:', webhook.url);

  // Wait for the webhook to be called.
  const req = await webhook;
  console.log('Received webhook request:', req);

  const body = await req.json();
  console.log('Webhook request body:', body);

  if (body.type === 'response.completed') {
    const text = await getOpenAIResponse(body.data.id);
    console.log('OpenAI response text:', text);
  }

  console.log('OpenAI webhook demo workflow completed');
}
```