Fix E2E
The environment variable `NATES_OPENAI_KEY` was removed from the
`uncurated-tests/example-workflow` Project.

The E2E tests were not directly relying on that env var, but the client
was failing to be initialized in the step function of the E2E tests even
though that specific workflow `withGetWebhook()` is not run in the
tests.

For now just defer the creation of the OpenAI client into the step
function body.