New typescript language server plugin (#272)
Started a new typescript language server plugin so we can provide useful developer hints/autocomplete/etc. in the IDE

This sets up the basic plugin and wires it up our own workbench apps. Once it's more mature and we've dogfooded, we can document and ship it within the `workflow` package, etc.

![CleanShot 2025-10-05 at 21.27.00@2x.png](https://app.graphite.dev/user-attachments/assets/bda3fe25-4743-4db2-b8ab-3745a93e9210.png)